Restructure component files
